# apispec.yaml
openapi: 3.0.0
info:
  description: |
    Track your calories everyday
  version: "1.0.0"
  title:  Daily Tracker Calories
servers:
  - url: 'https://localhost:8080/'
...
paths:
...
  '/food/{id}':
    get:
      tags:
        - food
      summary: Find Food by ID
      description: Returns a single Food
      operationId: getFoodById
      parameters:
        - name: id
          in: path
          description: ID of food to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Food'
        '400':
          description: Invalid ID supplied
        '404':
          description: Food not found
    put:
      tags:
        - food
      summary: Update an existing Food
      operationId: updateFood
      parameters:
        - name: id
          in: path
          description: ID of food to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID
        '404':
          description: Food not found
        '405':
          description: Validation exception
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Food'
        required: true
    delete:
      tags:
        - food
      summary: Delete Food
      description: This can only be done by the logged in user.
      operationId: deleteFood
      parameters:
        - name: id
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delete Success
        '400':
          description: Invalid Id
        '404':
          description: Food not found
  '/food/{name}':
    get:
      tags:
        - food
      summary: Find Food by Name
      description: Returns a single Food
      operationId: getFoodByName
      parameters:
        - name: name
          in: path
          description: Name of food to return
          required: true
          schema:
            type: string
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Food'
        '400':
          description: Invalid Name
        '404':
          description: Food not found
...
  /calorie:
    post:
      tags:
        - calories
      summary: Count Calorie
      description: This can only be done by the logged in user.
      operationId: createCalorie
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response_Calorie'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Calorie'
        description: Created user object
        required: true
  /calorie/save:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - calories
      summary: Save Calorie (Insert/Update)
      description: This can only be done by the logged in user.
      operationId: saveCalorie
      responses:
        '200':
          description: successful operation
          headers:
            Bearer-Token:
              description: JWT Auth
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response_Calorie'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Calorie'
        description: Created user object
        required: true
...
components:
  schemas:
...
    Food:
      type: object
      properties:
        id:
          type: integer
          format: uuid
          example: 15
        name:
          type: integer
          example: Bakso
        photo:
          type: string
          example: bakso.jpg
        calorie:
          type: integer
          example: 351
        created_at:
          type: string
          format: timestamp
          example: '124123413'
        updated_at:
          type: string
          format: timestamp
          example: '124123413'
...
    Calorie:
      properties:
        id:
          type: integer
          format: uuid
          example: 10
        weight:
          type: integer
          example: 50
        height:
          type: integer
          example: 165
        age:
          type: integer
          example: 25
        activity:
          type: integer
          example: 2
...
    Response_Calorie:
      properties:
        user_id:
          type: integer
          example: 1
        result_calorie:
          type: integer
          example: 2210
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT